shouldn't be exceptions just because boundaries are (currently) coterminous
As discussed SLDL/SLDU are still divisions even if they happen to be
coterminous at the moment.  They'll have different people associated
with them, and for historical purposes where they didn't always overlap
we need to keep in mind that there *are* upper & lower districts.

This commit leaves the NE redirects intact as those were indeed an error
and the redirects make sense.