add California assigned deferred senate districts
These were removed by 09c19b when we switched from an explicit list of
SLDU districts to a script to generate them.

I actually wrote a patch to census_places.py to add the assigned
districts, but this is such an exceptional case that just adding a
one-off list seems more sensible.

Note that CA state senate deferred areas are only applicable until the
newly elected state senators in even-numbered districts take office on
the first Monday of December 2014.