Add test that checks if updates have been compiled
The new test can be run with Bazel, from the root of the repository:

$ bazel test :all --test_output=errors

Any change to this repository that modifies sources for some country's
data (the stuff in identifiers/*/*.csv) should also modify the
corresponding compiled output (identifiers/country-??.csv), such that
they match.

This test runs the compiler, and compares its output to what's been
committed, ensuring they are equal. If this test fails, you need to
re-run the compiler for your country, and commit the result.