Use UTF-8 encoding regardless of locale
Python's open() function allows us to specify the encoding explicitly
with the `encoding=` kwarg. If unspecified, it uses the locale setting,
which may or may not be `UTF-8`.

The guidelines at
http://docs.opencivicdata.org/en/latest/proposals/0002.html specify that
all of our identifiers should use UTF-8. Therefore, this commit enforces
that the compile script and test read and write UTF-8.

This change was prompted when the compile_test was unable to decode the
`Ħ` character in `identifiers/country-mt.csv`, while running under the
Github test action, because in the test environment, the locale didn't
use UTF-8 by default. [Test output
here](https://github.com/jrunningen/ocd-division-ids/runs/1771915491?check_suite_focus=true).