Un-create countries that are territories of other countries
Civic identifiers are grouped a sovereign state. This commit removes
CSV files for ISO-3166 2-letter country codes where the territory
referred to falls under the sovereignty of another, according to the
table at https://en.wikipedia.org/wiki/List_of_ISO_3166_country_codes.

This is any country code with sovereignty not listed as "UN member
state," "UN observer state," or "disputed," and which did not previously
exist in the respository.