Populate country level OCD IDs for all countries in the world
For any country that did not yet have a CSV under identifiers, this
commit adds a CSV that defines the country code for that country.

This was created using the list of officially assigned code elements
from Wikipedia at
https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2#Officially_assigned_code_elements,
at the time of this commit.

Any countries that already existed in this repository were not touched.