Make the test ignore row ordering
Row ordering shouldn't be significant in this comparison. The compile.py
script will sort the rows before writing the output, which can be
surprising to someone who adds just one source file, copies it exactly
to identifiers/country-??.csv, but gets hit with a test failure because
the rows aren't sorted.